--- a/.babelrc
+++ b/.babelrc
@@ -1,8 +1,0 @@
-{
-    "presets": [
-        "next/babel"
-    ],
-    "plugins": [
-        "inline-react-svg"
-    ]
-}