--- a/package.json
+++ b/package.json
@@ -99,7 +99,6 @@
     "@types/react": "^18.0.9",
     "@types/supercluster": "^5.0.3",
     "autoprefixer": "^10.4.1",
-    "babel-plugin-inline-react-svg": "^2.0.1",
     "eslint-plugin-jest": "^25.3.4",
     "husky": "^7.0.4",
     "jest": "^27.5.1",