… src/assets/topography.svg
… yarn.lock