Trying out NextJS [Rust compiler](https://nextjs.org/docs/advanced-features/compiler).

While there isn't much improvement in `yarn build` command (since we don't prebuild that many pages), run time build ([server-side generation](https://nextjs.org/docs/basic-features/pages#static-generation-recommended)) is a significantly faster.  See SSG test below.

# Build test

### Before
```
...

├ ● /areas/[id] (ISR: 600 Seconds) (3994 ms)                 223 B           426 kB
├   ├ /areas/decc1251-4a67-52b9-b23f-3243e10e93d0 (1608 ms)
├   ├ /areas/78da26bc-cd94-5ac8-8e1c-815f7f30a28b (1399 ms)
├   └ /areas/bea6bf11-de53-5046-a5b4-b89217b7e9bc (781 ms)
├ λ /auth/signin                                             1.1 kB          297 kB
├ ● /climbs/[id] (493 ms)                                    440 kB          745 kB
├ ● /crag/[id]                                               2.21 kB         428 kB
├ ○ /export                                                  2.43 kB         298 kB
├ ○ /finder (350 ms)                                         19.8 kB         425 kB
├ ○ /tos                                                     5.79 kB         301 kB
└ ● /u/[...slug] (ISR: 60 Seconds) (4148 ms)                 8.48 kB         313 kB
    ├ /u/vietnguyen2 (2111 ms)
    └ /u/mario (1820 ms)
+ First Load JS shared by all                                103 kB
  ├ chunks/framework-1e63261378c6c882.js                     62.4 kB
  ├ chunks/main-87077a082bbb19fa.js                          26.7 kB
  ├ chunks/pages/_app-18ae906fd6b0cd38.js                    11.6 kB
  ├ chunks/webpack-c6ddeacc7a5b4840.js                       1.91 kB
  └ css/39ed1099362a56cf.css                                 14.5 kB

λ  (Server)  server-side renders at runtime (uses getInitialProps or getServerSideProps)
○  (Static)  automatically rendered as static HTML (uses no initial props)
●  (SSG)     automatically generated as static HTML + JSON (uses getStaticProps)
   (ISR)     incremental static regeneration (uses revalidate in getStaticProps)

✨  Done in 61.78s.

```

### After
```
...

├ ● /areas/[id] (ISR: 600 Seconds) (3361 ms)                 223 B           431 kB
├   ├ /areas/78da26bc-cd94-5ac8-8e1c-815f7f30a28b (1862 ms)
├   ├ /areas/bea6bf11-de53-5046-a5b4-b89217b7e9bc (772 ms)
├   └ /areas/decc1251-4a67-52b9-b23f-3243e10e93d0 (512 ms)
├ λ /auth/signin                                             1.24 kB         299 kB
├ ● /climbs/[id] (461 ms)                                    441 kB          749 kB
├ ● /crag/[id]                                               2.52 kB         433 kB
├ ○ /export                                                  2.38 kB         301 kB
├ ○ /finder                                                  20.3 kB         428 kB
├ ○ /tos                                                     5.78 kB         304 kB
└ ● /u/[...slug] (ISR: 60 Seconds) (4174 ms)                 9.74 kB         318 kB
    ├ /u/vietnguyen2 (2179 ms)
    └ /u/mario (1819 ms)
+ First Load JS shared by all                                105 kB
  ├ chunks/framework-1e63261378c6c882.js                     62.4 kB
  ├ chunks/main-3bc3decd973f58aa.js                          29.4 kB
  ├ chunks/pages/_app-661b7648934771af.js                    11.6 kB
  ├ chunks/webpack-b4a650beea72ee2e.js                       1.87 kB
  └ css/89b8d67293fe0bdf.css                                 14.6 kB

λ  (Server)  server-side renders at runtime (uses getInitialProps or getServerSideProps)
○  (Static)  automatically rendered as static HTML (uses no initial props)
●  (SSG)     automatically generated as static HTML + JSON (uses getStaticProps)
   (ISR)     incremental static regeneration (uses revalidate in getStaticProps)

✨  Done in 51.53s.

```

# Cold SSG requests

Test commands:
 ```
curl --output /dev/null   --silent --write-out '%{time_total}s\n' http://localhost:3000/climbs/1abda280-6f00-5729-bea8-13eb3f555cce

curl --output /dev/null   --silent --write-out '%{time_total}s\n' http://localhost:3000/areas/ab48aed5-2e8d-54bb-b099-6140fe1f098f
```

### Before
```
curl --output /dev/null   --silent --write-out '%{time_total}s\n' http://localhost:3000/climbs/1abda280-6f00-5729-bea8-13eb3f555cce

10.950470s

curl --output /dev/null   --silent --write-out '%{time_total}s\n' http://localhost:3000/areas/ab48aed5-2e8d-54bb-b099-6140fe1f098f

9.555152s
```


### After

```
curl --output /dev/null   --silent --write-out '%{time_total}s\n' http://localhost:3000b48aed5-2e8d-54bb-b099-6140fe1f098f
2.708992

curl --output /dev/null   --silent --write-out '%{time_total}s\n' http://localhost:3000/climbs/1abda280-6f00-5729-bea8-13eb3f555cce
4.402574s


```